Change the shape of the feature augmenting embeddings
Change the shape of the feature augmenting embeddings from
batch_size x n_embeddings x embedding_dim to
batch_size x n_choices x n_embeddings x embedding_dim. This change
allows for different embeddings to be applied to different choices,
enabling augmentation based on the answer option. Additionally,
change the shape of the features that come from the dataset reader
to pre-distribute the contexts so that it matches the shape of the
embeddings.

Finally, these code changes make the DistributeContextTransform no
longer necessary, so get rid of the code relating to it.