Restrict WinoGrande to only the physical portion
When the choices in WinoGrande are common nouns or noun phrases,
then an obvious feature augmentation strategy to try is simply
adding the ConceptNet features for the choice. When the options are
names, it's less clear how to augment the data. Therefore, we'll
begin experimenting by restricting only to the physical portion of
WinoGrande (questions where the options are common nouns or noun
phrases).