Add mixtures for producing learning curves
Add leave-one-task-out mixtures, where one of the tasks has a
reduced number of examples so that we can study how the multi-task
learning impacts performance across different dataset sizes. Because
there are 6 * 16 = 96 of these mixtures, the mixtures have to be
added in an iterative style (using for loops) rather than the
previous declarative style (writing everything out in the source
code).