Fix append mode bug in the conceptnet preparer
Opening a non-existent file in append mode from GCloud storage
throws an error when using GFile. In contrast, opening a
non-existent local file in append mode with GFile simply creates
the file. This commit uses write mode when the file doesn't exist
and append mode when it does in order to handle the GCloud use
case.