Fix a bug in table generation by correcting the size factor
In the experiments, the size (# examples) for the training set
would automatically be truncated to at most the full training set.
The analysis code would take the size parameter at face value.
Instead, bin/generate-tables.py should process the factors to
adjust for when the training set size was truncated.